requirements.txt
flask==2.3.3
flask-cors>=4.0.2
gunicorn>=22.0.0
google-generativeai==0.7.1
google-api-python-client==2.135.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
uvicorn
bytez
psutil
flask-limiter
urllib3>=2.0.0.post1
flask_sqlalchemy
requests
eventlet
